pyproject.toml
[project]
name = "perplexity-beta"
version = "0.1.0"
description = ""
authors = [
    {name = "valiante14",email = "joao.almeida14@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "langgraph (>=0.4.5,<0.5.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
